Add optional override feature in invert object utility helper
In utilities, `invertObject` can now take an optional 2nd parameter
which is an override function. The function takes the parameters
key and value and the override is only valid if it returns with an
array of 2 items. Otherwise, the default swapping behavior will
occur as usual.

The motivation for adding this feature is because React's
SVGDOMPropertyConfig has properrty keys with values of 0. As a
result, the override will come in handy when dealing with those
properties.

Add utilities tests to verify the new feature in the helper.