Update attributes to props helper to handle SVG props
The property config originally only had the HTML DOM property
config. Updated the property config with the SVG property
config as well.

Then update `attributesToProps` to try setting both HTML and SVG
properties when iterating through the attributes.

The only difference between the HTML and SVG config is that
the HTML property keys are lowercased whereas the SVG property
keys are not because there are certain keys that are camel cased.

Add tests to check that the SVG attributes and properties are
handled correctly by the helper.

React's SVGDOMPropertyConfig source:
https://github.com/facebook/react/blob/master/src/renderers/dom/shared/SVGDOMPropertyConfig.js