Add optional replace override to parser
Update `domToReact` and `htmlToReact` to have a second and optional
argument which is an object with a `replace` method.

The method will give the user access to the DOM node object and if
a valid React element is returned, then it will override the node.

Thus, the user can set the logic or condition in which a node is
replaced with a custom React element.

Add tests to check that the option works as expected.
Tidy up and organize HTML to React tests.