Create test helper that does deepEqual for circular references
Since the objects returned by `htmlparser2.parseDOM` have circular
references, this makes `assert.deepEqual` testing difficult.

As a result, a test helper was created to handle this issue.
`deepEqualCircular` walks through the objects and taints the
visited ones to prevent the call stack from being overwhelmed.

A unit test was also created to check that the helper logic is
consistent.