Remove `key` parameter from `replace` and use `React.cloneElement`
The previous implementation was to pass the `key` as the 2nd
parameter in `replace` method and force the user to set the
"key" prop in their custom React element.

This creates further complexity for the end-user.

As a result, use `React.cloneElement` which does the same thing
without additional work from the user.

`React.cloneElement` takes the React element and merges the
"key" prop taken from the array index with the original
element's props when applicable.

https://facebook.github.io/react/docs/top-level-api.html#react.cloneelement