Fix bug since DOMPropertyConfigs are moved to `react-dom` in v15.4
Both `HTMLDOMPropertyConfig.js` and `SVGDOMPropertyConfig.js` used
to be in `react/lib/` directory but since v15.4.x, the configs
have been moved to `react-dom/lib/`.

Created a module that tried to require either `react-dom/lib/`
or `react/lib/` (fallback).

A more permanent (less error-prone) solution may be to keep a
hardcopy of both configs in the repository (minimize coupling and
risk of another change from breaking the package).

Fixes #29