fix(attributes-to-props): handle attr named after Object properties
Since the property config is a POJO (Plain Old JavaScript Object),
this means the Object properties/methods are accessible.

Thus, use `hasOwnProperty` to check if the attribute name is on the
config map rather than on the Object prototype chain.

Handle attributes that share the same name as custom attributes.