build(index): fix rollup error of mixing named and default exports
```
(!) Mixing named and default exports
Consumers of your bundle will have to use bundle['default'] to access the default export, which may not be what you want. Use `output.exports: 'named'` to disable this warning
```

This fixes the UMD bundle (in `dist`) as the library can only be
accessed via the default export.